requirements.txt
fastapi==0.101.0
uvicorn[standard]==0.23.1
python-dotenv==1.0.0
sentence-transformers==2.2.2
faiss-cpu==1.12.0
numpy==1.26.2
requests==2.31.0